+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+### Changed
+### Fixed
+- bug #212: now it is possible to use several external commands in a pipeline
+
 ## [0.0.36] - 2020-01-30
 
 ### Added